Fold UrlFormParser into UrlForm, restrict internals visibility
Also uppercase sandbox constant names.
