+++ modules/core/src/smithy4s/http/UrlForm.scala
@@ -27,11 +27,15 @@ import smithy4s.http.internals.UrlFormDataEncoderSchemaVisitor
 import smithy4s.schema.CachedSchemaCompiler
 import smithy4s.schema.Schema
 
+import java.io.UnsupportedEncodingException
+import java.net.URLDecoder
 import java.net.URLEncoder
+import java.nio.CharBuffer
 import java.nio.charset.StandardCharsets
+import scala.collection.immutable.BitSet
 import scala.collection.mutable
 
-final case class UrlForm(values: List[UrlForm.FormData]) {
+private[smithy4s] final case class UrlForm(values: List[UrlForm.FormData]) {
 
   def render: String = {
     val builder = new mutable.StringBuilder
@@ -76,7 +80,84 @@ private[smithy4s] object UrlForm {
     }
   }
 
-  trait Decoder[A] {
+  // This is based on http4s' own equivalent, but simplified for our use case.
+  def parse(urlFormString: String): Either[UrlFormDecodeError, UrlForm] = {
+    val inputBuffer = CharBuffer.wrap(urlFormString)
+    val encodedTermBuilder = new StringBuilder(capacity = 32)
+    val outputBuilder = List.newBuilder[UrlForm.FormData]
+
+    var state: State = Key
+    var error: UrlFormDecodeError = null
+    var key: String = null
+
+    def endPair(): Unit = {
+      appendPair()
+      state = Key
+    }
+
+    def appendPair(): Unit = if (state == Key) {
+      outputBuilder += UrlForm.FormData(
+        PayloadPath.parse(decodeTerm(encodedTermBuilder.result())),
+        maybeValue = None
+      )
+      encodedTermBuilder.clear()
+    } else {
+      outputBuilder += UrlForm.FormData(
+        PayloadPath.parse(decodeTerm(key)),
+        Some(decodeTerm(encodedTermBuilder.result()))
+      )
+      key = null
+      encodedTermBuilder.clear()
+    }
+
+    def decodeTerm(str: String): String =
+      try URLDecoder.decode(str, StandardCharsets.UTF_8.name())
+      catch {
+        case _: UnsupportedEncodingException => ""
+      }
+
+    while (error == null && inputBuffer.hasRemaining)
+      inputBuffer.get() match {
+        case '&' => endPair()
+        case '=' =>
+          if (state == Value) encodedTermBuilder.append('=')
+          else {
+            state = Value
+            key = encodedTermBuilder.result()
+            encodedTermBuilder.clear()
+          }
+        case char if QueryChars.contains(char.toInt) =>
+          encodedTermBuilder.append(char)
+        case char =>
+          error = UrlFormDecodeError(
+            PayloadPath.root,
+            s"Invalid char while splitting key/value pairs: '$char'"
+          )
+      }
+
+    if (error != null) Left(error)
+    else {
+      appendPair()
+      Right(UrlForm(outputBuilder.result()))
+    }
+  }
+
+  private sealed trait State
+  private case object Key extends State
+  private case object Value extends State
+
+  // These are the characters that are allowed unquoted within a query string as
+  // defined in https://datatracker.ietf.org/doc/html/rfc3986#appendix-A.
+  private val QueryChars: BitSet = BitSet(
+    (Pchar ++ "/?".toSet - '&' - '=').map(_.toInt).toSeq: _*
+  )
+
+  private def Pchar = Unreserved ++ SubDelims ++ ":@%".toSet
+  private def Unreserved = "-._~".toSet ++ AlphaNum
+  private def AlphaNum = (('a' to 'z') ++ ('A' to 'Z') ++ ('0' to '9')).toSet
+  private def SubDelims = "!$&'()*+,;=".toSet
+
+  private[smithy4s] trait Decoder[A] {
     def decode(urlForm: UrlForm): Either[UrlFormDecodeError, A]
   }
