+++ modules/core/src/smithy4s/http/internals/UrlFormDataEncoder.scala
@@ -22,7 +22,7 @@ import smithy4s.capability.EncoderK
 import smithy4s.codecs.PayloadPath
 import smithy4s.http.UrlForm
 
-private[smithy4s] trait UrlFormDataEncoder[-A] { self =>
+private[http] trait UrlFormDataEncoder[-A] { self =>
 
   def encode(value: A): List[UrlForm.FormData]
 
@@ -34,7 +34,7 @@ private[smithy4s] trait UrlFormDataEncoder[-A] { self =>
 
 }
 
-object UrlFormDataEncoder {
+private[internals] object UrlFormDataEncoder {
 
   implicit val urlFormDataEncoderK
       : EncoderK[UrlFormDataEncoder, List[UrlForm.FormData]] =
